Consolidate cloud backup store
