+++ rust/src/manager/cloud_backup_manager/ops/test_support.rs
@@ -717,8 +717,8 @@ pub(crate) fn persist_dirty_blob_state(wallet_id: WalletId) {
 }
 
 fn mutate_wallet_and_persist_dirty(wallet_id: WalletId) {
-    let mut wallet = CloudBackupWalletStore::global()
-        .all()
+    let mut wallet = CloudBackupStore::global()
+        .all_wallets()
         .unwrap()
         .into_iter()
         .find(|wallet| wallet.id == wallet_id)
