+++ rust/src/manager/cloud_backup_manager/wallets/payload.rs
@@ -9,10 +9,10 @@ use serde::Serialize;
 use sha2::{Digest, Sha256};
 use tracing::warn;
 
-use super::{CloudBackupError, LocalWalletMode, MAX_CLOUD_LABELS_SIZE, PreparedWalletBackup};
+use super::{MAX_CLOUD_LABELS_SIZE, PreparedWalletBackup};
 use crate::backup::model::DescriptorPair as LocalDescriptorPair;
 use crate::label_manager::LabelManager;
-use crate::manager::cloud_backup_manager::LocalWalletSecret;
+use crate::manager::cloud_backup_manager::{CloudBackupError, LocalWalletMode, LocalWalletSecret};
 use crate::mnemonic::MnemonicExt as _;
 use crate::wallet::{
     WalletAddressType,
